Don't assert on AWS_PROFILE anymore
There are a number of ways to choose your ARN, and AWS_PROFILE is only
one of them. With the removal of the stderr redirect, users will get
error messages from sops if they don't have the ability to assume an AWS
role, so just rely on that for accurate error reporting.